Add Report Export Functionality (#35)
* add a new optional param to refresh funcs

this will not break existing references to the functions and will ensure that the new behaviour will be off unless explicitly enabled at function call

* create a new report generator function

The new system report generator function will produce 2 strings (1 md & 1 html) which can in the future be connected to a system report export system, two new modules have been added to the import list to enable this function to work (markdown & datetime).

* add a temporary way to test the report gen

prints the output of the report generator to the terminal when the refresh button is ran, this will be removed when an actual system to export the reports has been implemented

* create new variables to hold

this will prevent a situation where two string literals are being fed into two different functions which would inevitably become an issue if these literals are not updated at the same time, using a single var for both eliminates this risk

* missed from previous commit

oops, my bad

* minor formatting fix

this very minor inconsistancy bugged me abit...

* confused as to the purpose of this var

im unsure if its best practice to record the issue in that manner but i would have forgotten otherwise

* data export on motherboard func

whenever data has been collected and written to the window a check is done to see if "export_data" is enabled, if so it will add that data to a string which will be exported to a report, at the end of the function - if export_data is enabled - it will then return this string

* data export on cpu func

see motherboard data export commit desc for details

* data export on network func

see motherboard data export commit desc for details

* data export on pci func

see motherboard data export commit desc for details (also the variable "first_part" is created (ln 423) but never displayed in the window, i have included it in the report but if this should not be done for whatever reason please let me know

* data export on memory func

* data export on disk func

* data export on kernel func

* data export on system func

* add two new buttons for exporting reports

* adjust window sizing so all buttons show on launch

* button functionality working now

* remove miliseconds from the datetime

* more output formatting

* implement file saving system!

everything works™, some formatting and performance issues likely still need addressing but the core functionality is all there

* remove un-needed debugging print statement

* delete the cached report once done

such a big global var will use a fair bit of ram, once the file is saved we can free up this space again

* add to dev list! :D

* fix global var issue

also remove redundant parsing of the report var to the file_save_dialog function (it can't pass it to file save response anyway)

* minor performance optimisation

move the html conversion to the html only function so we don't waste time converting to html for an md export

* fix error thrown when export canceled

previously an error was thrown by opening the save file dialog for export and then closing it, this is no longer the case

* fix name of hml export func

was uppercase (inconsistent with md export func name)

* add pdf export function husk

currently just does the same as the md export but will allow for testing of the md to pdf functionality and button testing

* nasty import problems

the modules on PyPI have their names `-` separated but as that is invalid py import syntax they must be imported `_` separated, this then means meson fails to find the module as it is apparently only distributed with the `-` separated name

* meson seems to refuse to install some modules

it installs io perfectly fine but refuses to install markdown-it, markdown-pdf or xhtml2pdf? I am going to assume im doing something wrong and have left the code commented out so everything else is still functional

* fix formatting as per suggestions

now the reports are *far* more compact, the html is also actually readable now thanks to this!

* get rid of deprecated vars

* fix the import/build errors

many thanks to Nokse22 for the jsons and general pointers

Co-Authored-By: Nokse <44558032+Nokse22@users.noreply.github.com>

* switch from markdown to markdown_it module

* working PDF export func

tested and correctly converts the fed HTML fed into it into an identically formatted pdf

* added drop down menu, it won't work though?

* put all export options into a single button

button brings up a drop down box containing the 3 options, currently the keybinds are set to ctrl+ the first letter of the file type, these need to be added to the keyboard shortcut window at some stage

* missed tooltip for new menu button

* implement formatting suggestions from nokse22

Co-Authored-By: Nokse <44558032+Nokse22@users.noreply.github.com>

* added keyboard shortcuts to the list

---------

Co-authored-by: Nokse <44558032+Nokse22@users.noreply.github.com>